1.更新編輯後台會員 2.後台會員刪除改為假刪除 3.前台訂單可用人數及後台API目前可用人數改為總秒數判斷 4.更新後台訂單成本計算

+++ HeO.Models/Feedbackrecord.cs
@@ -20,7 +20,7 @@ namespace HeO.Models
         public string Hdzaccount { get; set; }
         public int Total { get; set; }
         public int Money { get; set; }
-        public int Remains { get; set; }
+        public double Remains { get; set; }
         public int Cash { get; set; }
         public string Bankaccount { get; set; }
         public string Remark { get; set; }
